requirements.txt
Django==1.9.6
django-markup==1.1
djangorestframework==3.3.3
djangorestframework-apidoc==0.0.1
djangorestframework-jsonp==1.0.2
djangorestframework-recursive==0.1.1
psycopg2==2.6.1
six==1.10.0
